Store instruction arguments as usize
While instruction arguments are read as a u16 in the bytecode parser
they are now stored as a usize. This removes the need for casting u16
values to usize values during runtime, speeding up the interpreter a
bit.

+++ vm/src/vm/instructions/test.rs
@@ -46,7 +46,9 @@ pub fn setup() -> (Machine, Block, RcProcess) {
 }
 
 /// Creates a new instruction.
-pub fn new_instruction(ins_type: InstructionType, args: Vec<u16>) -> Instruction {
+pub fn new_instruction(ins_type: InstructionType,
+                       args: Vec<usize>)
+                       -> Instruction {
     Instruction::new(ins_type, args, 1)
 }
